test: Abstract common functions from script break ups in test_bgp_auth.py
